flexbox
- Tried to move most of the css to "style.css" so it isn't spread out
between <style="">, <style> and style.css. Some of it got lost along the
way, but it is easy to recreate. I added the backup files for that
purpose.
- Made the Search View and the Details View use flexbox instead of a
static css. This means that the window is resizeable without everything
breaking.
- Also, I left the borders in there so you easily can see what to style
when you want to style something.
- Also, fixed the drag and drop so the dragged object becomes a clone in
front of everything instead of a sibling element.